refactor(linter): extract json parsing out of parser.rs to satisfy file length limits
- Extract JSON handling from `utils/parser.rs` to `utils/json.rs`
- Expand `ast_linter_file_length` test to enforce 200-line limit across the entire `katana-linter/src` tree
- Confirmed all linter files are now compliant with the length constraint
