+++ crates/katana-linter/src/utils/mod.rs
@@ -1,10 +1,12 @@
 /// File system collection utilities for the AST linter.
 pub mod file_collector;
+pub mod json;
 /// Source code and JSON parsing utilities for the AST linter.
 pub mod parser;
 /// Diagnostics and report generation utilities.
 pub mod reporter;
 
 pub use file_collector::*;
+pub use json::*;
 pub use parser::*;
 pub use reporter::*;
